Added restart and improved the start and stop functionality.

+++ cli/scripts/linux/remove.sh
@@ -1,3 +1,5 @@
 #!/bin/bash
 
-docker rm -f <%= name %>
+if [ -n $(docker ps -aq -f name="<%= name %>") ]; then
+  docker rm -f <%= name %>
+fi
